refactor: move AgendaItem to domain/model
