+++ androidApp/src/main/java/com/example/tasky/android/agenda/components/details/DetailsDateTimeSection.kt
@@ -28,10 +28,10 @@ import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
-import com.example.tasky.agenda.data.model.AgendaItem
-import com.example.tasky.agenda.data.model.Event
-import com.example.tasky.agenda.data.model.Reminder
-import com.example.tasky.agenda.data.model.Task
+import com.example.tasky.agenda.domain.model.AgendaItem
+import com.example.tasky.agenda.domain.model.Event
+import com.example.tasky.agenda.domain.model.Reminder
+import com.example.tasky.agenda.domain.model.Task
 import com.example.tasky.android.R
 import com.example.tasky.android.theme.Black
 import com.example.tasky.android.theme.MyApplicationTheme
