+++ androidApp/src/main/java/com/example/tasky/android/agenda/components/details/DetailsPhotoSection.kt
@@ -33,7 +33,7 @@ import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import coil3.compose.AsyncImage
-import com.example.tasky.agenda.data.model.Photo
+import com.example.tasky.agenda.domain.model.Photo
 import com.example.tasky.android.R
 import com.example.tasky.android.theme.Black
 import com.example.tasky.android.theme.Gray
