+++ shared/src/commonMain/kotlin/com/example/tasky/agenda/data/AgendaDataSource.kt
@@ -4,16 +4,16 @@ import com.example.tasky.agenda.data.model.Agenda
 import com.example.tasky.agenda.data.model.CreateEventBody
 import com.example.tasky.agenda.data.model.CreateReminderBody
 import com.example.tasky.agenda.data.model.CreateTaskBody
-import com.example.tasky.agenda.data.model.Event
 import com.example.tasky.agenda.data.model.EventPath
 import com.example.tasky.agenda.data.model.GetAgendaResponse
-import com.example.tasky.agenda.data.model.Reminder
 import com.example.tasky.agenda.data.model.ReminderPath
-import com.example.tasky.agenda.data.model.Task
 import com.example.tasky.agenda.data.model.TaskPath
 import com.example.tasky.agenda.data.model.UpdateEventBody
 import com.example.tasky.agenda.data.model.UpdateReminderBody
 import com.example.tasky.agenda.data.model.UpdateTaskBody
+import com.example.tasky.agenda.domain.model.Event
+import com.example.tasky.agenda.domain.model.Reminder
+import com.example.tasky.agenda.domain.model.Task
 import com.example.tasky.common.manager.HttpManager
 import com.example.tasky.common.model.BaseError
 import com.example.tasky.common.model.ResultWrapper
