+++ shared/src/commonMain/kotlin/com/example/tasky/agenda/data/model/Agenda.kt
@@ -1,5 +1,8 @@
 package com.example.tasky.agenda.data.model
 
+import com.example.tasky.agenda.domain.model.Event
+import com.example.tasky.agenda.domain.model.Reminder
+import com.example.tasky.agenda.domain.model.Task
 import io.ktor.resources.Resource
 import kotlinx.serialization.Serializable
