+++ shared/src/commonMain/kotlin/com/example/tasky/agenda/data/model/Event.kt
@@ -1,5 +1,6 @@
 package com.example.tasky.agenda.data.model
 
+import com.example.tasky.agenda.domain.model.Event
 import io.ktor.resources.Resource
 import kotlinx.serialization.Serializable
