+++ shared/src/commonMain/kotlin/com/example/tasky/agenda/data/model/Task.kt
@@ -1,5 +1,6 @@
 package com.example.tasky.agenda.data.model
 
+import com.example.tasky.agenda.domain.model.Task
 import io.ktor.resources.Resource
 
 @Resource("/task")
